refactor: FilenameExtension 이동
